move bom components in common components
